Extract regular export back out of resolver manager
